# tpl/advanced/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "描述信息",
        "title": "prajna 项目",
        "contact": {
            "name": "spruce"
        },
        "version": "1.0"
    },
    "host": "localhost:9009",
    "paths": {
        "/api/auth/info": {
            "get": {
                "description": "当前登录信息(需授权),从token中解析userid和roleid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Info",
                "responses": {}
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "demo登录(账号:name,密码:password),返回双token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号(demo名称)",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Login 成功\",\"data\":{\"accessToken\":\"\",\"refreshToken\":\"\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "登出,废弃当前token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Logout",
                "responses": {}
            }
        },
        "/api/auth/refresh": {
            "post": {
                "description": "刷新双token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问token",
                        "name": "accessToken",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "刷新token",
                        "name": "refreshToken",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/auth/ws-ticket": {
            "post": {
                "description": "申请 WebSocket 一次性升级票据(需授权),用 ticket 连接 WS,避免长 token 进 URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "WSTicket",
                "responses": {}
            }
        },
        "/api/demo/:id": {
            "get": {
                "description": "demo详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "demoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/demo/create": {
            "post": {
                "description": "写一个demo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "demoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/demo/excel": {
            "get": {
                "description": "http+mysql 全链路流式导出 Excel，浏览器直接下载",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Excel",
                "responses": {}
            }
        },
        "/api/demo/export": {
            "get": {
                "description": "http+mysql 全链路流式导出 demo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Export",
                "responses": {}
            }
        },
        "/api/demo/list": {
            "get": {
                "description": "根据条件查询demo列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "List",
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/demo/sse": {
            "get": {
                "description": "Server-Sent Events demo",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "api"
                ],
                "summary": "SSE",
                "responses": {}
            }
        },
        "/api/demo/update": {
            "post": {
                "description": "更新一个demo,有事务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Update",
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/demo/delete": {
            "post": {
                "description": "删除一个demo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "demoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/demo/softdelete": {
            "post": {
                "description": "逻辑删除一个demo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "SoftDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "demoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"demo-Demo成功\",\"data\":{\"id\":100,\"name\":\"我是demo名称\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":50000,\"message\":\"demo查询失败\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}